Add files via upload
Band Name Generator Game
This fun and interactive game allows you to create your own band name! Simply follow the prompts, and the generator will come up with a unique name based on your inputs. The game is in Spanish, so it's a great opportunity to practice the language while having fun! Enjoy and let your creativity flow!